[Refactor] 라우터 별도 파일로 분리
- BrowserRouter 생성 코드를 별도 파일로 분리
