+++ src/App.js
@@ -1,28 +1,7 @@
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { RouterProvider } from "react-router-dom";
 import { AuthContextProvider } from "./context/AuthContext";
-import Root from "./pages/Root";
-import Home from "./pages/Home";
-import ErrorPage from "./pages/ErrorPage";
-import Cart from "./pages/Cart";
-import Admin from "./pages/Admin";
-import Products from "./pages/Products";
-import ProductDetail from "./pages/ProductDetail";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root />,
-    errorElement: <ErrorPage />,
-    children: [
-      { index: true, element: <Home /> },
-      { path: "/cart", element: <Cart /> },
-      { path: "/admin", element: <Admin /> },
-      { path: "/products", element: <Products /> },
-      { path: "/products/:id", element: <ProductDetail /> },
-    ],
-  },
-]);
+import router from "./routes/router";
 
 const queryClient = new QueryClient();
